requirements.txt
fastapi
pydantic
uvicorn
xgboost
scikit-learn
requests
sqlalchemy 
pandas 
psycopg2-binary 
python-dotenv
jupyter
optuna
dvc_s3
pytest
prefect
dvclive
evidently
boto3